+++ cellprofiler/modules/tests/test_tile.py
@@ -175,7 +175,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
     
     def test_02_01_manual_rows_and_columns(self):
         np.random.seed(0)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -205,7 +205,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
             
     def test_02_02_automatic_rows(self):
         np.random.seed(1)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -235,7 +235,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
             
     def test_02_03_automatic_columns(self):
         np.random.seed(2)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -265,7 +265,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
             
     def test_02_04_automatic_rows_and_columns(self):
         np.random.seed(3)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -294,7 +294,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
             
     def test_02_05_color(self):
         np.random.seed(4)
-        images = [np.random.uniform(size=(20,10,3)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10,3)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -324,7 +324,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
 
     def test_02_06_columns_first(self):
         np.random.seed(5)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -354,7 +354,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
 
     def test_02_07_top_right(self):
         np.random.seed(0)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -385,7 +385,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
 
     def test_02_08_bottom_left(self):
         np.random.seed(8)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -416,7 +416,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
 
     def test_02_09_bottom_right(self):
         np.random.seed(9)
-        images = [np.random.uniform(size=(20,10)) for i in range(96)]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32) for i in range(96)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -447,10 +447,10 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
             
     def test_02_10_different_sizes(self):
         np.random.seed(10)
-        images = [np.random.uniform(size=(20,10)),
-                  np.random.uniform(size=(10,20)),
-                  np.random.uniform(size=(40, 5)),
-                  np.random.uniform(size=(40,20))]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32),
+                  np.random.uniform(size=(10,20)).astype(np.float32),
+                  np.random.uniform(size=(40, 5)).astype(np.float32),
+                  np.random.uniform(size=(40,20)).astype(np.float32)]
         workspace, module = self.make_tile_workspace(images)
         self.assertTrue(isinstance(module, T.Tile))
         self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -512,7 +512,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
     def test_03_01_some_images(self):
         np.random.seed(31)
         for i in range(1,5):
-            images = [np.random.uniform(size=(20,10)) for ii in range(i)]
+            images = [np.random.uniform(size=(20,10)).astype(np.float32) for ii in range(i)]
             workspace, module = self.make_place_workspace(images)
             self.assertTrue(isinstance(module, T.Tile))
             self.assertTrue(isinstance(workspace, cpw.Workspace))
@@ -527,7 +527,7 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
     def test_03_02_mix_color_bw(self):
         np.random.seed(32)
         for color in range(3):
-            images = [np.random.uniform(size=(20,10,3) if i == color else (20,10))
+            images = [np.random.uniform(size=(20,10,3) if i == color else (20,10)).astype(np.float32)
                       for i in range(3)]
             workspace, module = self.make_place_workspace(images)
             module.run(workspace)
@@ -544,10 +544,10 @@ Tile:[module_num:1|svn_version:\'9034\'|variable_revision_number:1|show_window:T
             
     def test_03_03_different_sizes(self):
         np.random.seed(33)
-        images = [np.random.uniform(size=(20,10)),
-                  np.random.uniform(size=(10,20)),
-                  np.random.uniform(size=(40, 5)),
-                  np.random.uniform(size=(40,20))]
+        images = [np.random.uniform(size=(20,10)).astype(np.float32),
+                  np.random.uniform(size=(10,20)).astype(np.float32),
+                  np.random.uniform(size=(40, 5)).astype(np.float32),
+                  np.random.uniform(size=(40,20)).astype(np.float32)]
         workspace, module = self.make_place_workspace(images)
         module.run(workspace)
         image = workspace.image_set.get_image(OUTPUT_IMAGE_NAME)
